pyproject.toml
[tool.poetry]
name = "python-workspace"
version = "1.0.0"
description = ""
authors = ["James Neale <42097394+james9226@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
pandas = "^1.5.0"
numpy = "^1.23.3"
seaborn = "^0.12.0"
streamlit = "^1.13.0"
protobuf = "3.19.5"
Pillow = "^9.3.0"
google-cloud-firestore = "^2.10.0"
google-auth = "^2.16.2"
bcrypt = "^4.0.1"
extra-streamlit-components = "^0.1.56"
PyJWT = "^2.6.0"

[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
mkdocs-material = "^9.0.11"
mkdocs-table-reader-plugin = "^2.0"
mkdocs-glightbox = "^0.3.1"
black = "^22.10.0"
mkdocs = "^1.4.0"
flake8 = "^5.0.4"
mypy = "^1.1.1"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = [
    "httplib2",
    "google.oauth2",
    "pandas",
    "extra_streamlit_components",
    "google.cloud"
]
ignore_missing_imports = true
